+++ src/commands/command/codecs/create-table.ts
@@ -1,7 +1,0 @@
-import { CreateTableInput } from "../inputs/create-table.js";
-import { CreateTableOutput } from "../outputs/create-table.js";
-
-export const CreateTableCodec = {
-  encode: CreateTableInput.encode,
-  decode: CreateTableOutput.decode,
-};