+++ src/commands/command/codecs/put-item.ts
@@ -1,7 +1,0 @@
-import { PutItemInput } from "../inputs/put-item.js";
-import { PutItemOutput } from "../outputs/put-item.js";
-
-export const PutItemCodec = {
-  encode: PutItemInput.encode,
-  decode: PutItemOutput.decode,
-};