+++ src/commands/command/outputs/index.ts
@@ -1,8 +1,0 @@
-export type { CreateTableOutput } from "./create-table.js";
-export type { DeleteItemOutput } from "./delete-item.js";
-export type { DeleteTableOutput } from "./delete-table.js";
-export type { GetItemOutput } from "./get-item.js";
-export type { PutItemOutput } from "./put-item.js";
-export type { QueryOutput } from "./query.js";
-export type { UpdateItemOutput } from "./update-item.js";
-export type { UpdateTimeToLiveOutput } from "./update-time-to-live.js";