+++ src/commands/expressions/index.ts
@@ -2,10 +2,3 @@ export * from "./condition/index.js";
 export * from "./key-condition.js";
 export * from "./operands/index.js";
 export * from "./update/index.js";
-
-/*
-TODO: In the expression interfaces, a slight QoL improvement would be to assume that
-strings are attribute names by default, and other types are values by default. The user
-always has the option to be explicit. For example, he could write:
-`where(attribute("name")).eq(value(42))` or `where("name").eq(42)`.
- */