+++ src/commands/index.ts
@@ -1,6 +1,38 @@
-export * from "./command/index.js";
+export * from "./attributes/index.js";
+export type { CommandInput, CommandOutput } from "./base.js";
+export type { DynamoDbClientCommand } from "./command.js";
+export { CreateTable } from "./create-table.js";
+export { DeleteItem } from "./delete-item.js";
+export { DeleteTable } from "./delete-table.js";
 export * from "./expressions/index.js";
+export { GetItem } from "./get-item.js";
 export type {
-  ConditionCheckParams,
-  WriteTransactionParams,
-} from "./write-transaction.js";
+  ConditionCheckInput,
+  CreateTableInput,
+  DeleteItemInput,
+  DeleteItemReturnValue,
+  DeleteTableInput,
+  GetItemInput,
+  PutItemInput,
+  QueryInput,
+  UpdateItemInput,
+  UpdateTimeToLiveInput,
+  WriteTransactionInput,
+  WriteTransactionWrite,
+} from "./inputs/index.js";
+export type {
+  CreateTableOutput,
+  DeleteItemOutput,
+  DeleteTableOutput,
+  GetItemOutput,
+  PutItemOutput,
+  QueryOutput,
+  UpdateItemOutput,
+  UpdateTimeToLiveOutput,
+  WriteTransactionOutput,
+} from "./outputs/index.js";
+export { PutItem } from "./put-item.js";
+export { Query } from "./query.js";
+export { UpdateItem } from "./update-item.js";
+export { UpdateTimeToLive } from "./update-time-to-live.js";
+export { WriteTransaction } from "./write-transaction.js";