+++ src/commands/types.ts
@@ -1,4 +1,0 @@
-export interface Command<Input, AwsCommand> {
-  toAwsCommandInput(): Input;
-  toAwsCommand(): AwsCommand;
-}