+++ src/index.ts
@@ -2,6 +2,6 @@
 export * from "./client.js";
 export * from "./commands/index.js";
 export * from "./error.js";
+export * from "./lib.js";
 export * from "./native-client-errors.js";
 export * from "./types.js";
-export * from "./utils.js";